requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0,<2.1
psycopg2-binary
alembic
pydantic>=2.0
passlib[bcrypt]
python-jose[cryptography]
aiokafka
confluent-kafka
